SDIT-824 Move all functional tests together and delete all unit tests
