+++ src/test/kotlin/uk/gov/justice/digital/hmpps/gradle/functional/configmanagers/BaseConfigManagerGradleTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.gradle.configmanagers
+package uk.gov.justice.digital.hmpps.gradle.functional.configmanagers
 
 import org.assertj.core.api.Assertions.assertThat
 import org.gradle.testkit.runner.TaskOutcome
@@ -12,62 +12,62 @@ import uk.gov.justice.digital.hmpps.gradle.functional.makeProject
 import java.io.File
 import java.nio.file.Files
 
-class BaseConfigManagerTrivyignoreTest : GradleBuildTest() {
+class BaseConfigManagerGradleTest : GradleBuildTest() {
   @ParameterizedTest
   @MethodSource("defaultProjectDetails")
-  fun `The trivy ignore file is copied into the project`(projectDetails: ProjectDetails) {
+  fun `The sonar properties file is copied into the project`(projectDetails: ProjectDetails) {
     makeProject(projectDetails)
 
     val result = buildProject(Companion.projectDir, "tasks")
     assertThat(result.task(":tasks")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
 
-    val trivyFile = findFile(projectDir, ".trivyignore")
-    assertThat(trivyFile).exists()
+    val sonarFile = findFile(projectDir, "sonar-project.properties")
+    assertThat(sonarFile).exists()
   }
 
   @ParameterizedTest
   @MethodSource("defaultProjectDetails")
-  fun `The trivy ignore file is not copied into the project`(projectDetails: ProjectDetails) {
+  fun `The sonar properties file is not copied into the project`(projectDetails: ProjectDetails) {
     makeProject(projectDetails)
 
-    val trivyScript =
+    val sonarScript =
       """
 # some coverage rules
-trivy.coverage.exclusions=**/*.java,**/*.kt
+sonar.coverage.exclusions=**/*.java,**/*.kt
       """.trimIndent()
-    makeTrivyFile(trivyScript)
+    makeSonarFile(sonarScript)
 
     val result = buildProject(Companion.projectDir, "tasks")
     assertThat(result.task(":tasks")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
 
-    val trivyFile = findFile(projectDir, ".trivyignore")
-    assertThat(trivyFile).exists()
-    val firstLine = trivyFile.useLines { it.firstOrNull() }
+    val sonarFile = findFile(projectDir, "sonar-project.properties")
+    assertThat(sonarFile).exists()
+    val firstLine = sonarFile.useLines { it.firstOrNull() }
     assertThat(firstLine).startsWith("# some coverage rules")
   }
 
   @ParameterizedTest
   @MethodSource("defaultProjectDetails")
-  fun `The trivy ignore file is overwritten in the project if WARNING exists`(projectDetails: ProjectDetails) {
+  fun `The sonar properties file is overwritten in the project if WARNING exists`(projectDetails: ProjectDetails) {
     makeProject(projectDetails)
-    val trivyScript =
+    val sonarScript =
       """
 # WARNING - contents will be overwritten
-trivy.coverage.exclusions=**/*.java,**/*.kt
+sonar.coverage.exclusions=**/*.java,**/*.kt
       """.trimIndent()
-    makeTrivyFile(trivyScript)
+    makeSonarFile(sonarScript)
 
     val result = buildProject(Companion.projectDir, "tasks")
     assertThat(result.task(":tasks")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
 
-    val trivyFile = findFile(projectDir, ".trivyignore")
-    assertThat(trivyFile).exists()
-    val firstLine = trivyFile.useLines { it.firstOrNull() }
+    val sonarFile = findFile(projectDir, "sonar-project.properties")
+    assertThat(sonarFile).exists()
+    val firstLine = sonarFile.useLines { it.firstOrNull() }
     assertThat(firstLine).startsWith("# WARNING - THIS FILE WAS GENERATED")
   }
 
-  private fun makeTrivyFile(trivyScript: String) {
-    val trivyFile = File(projectDir, ".trivyignore")
-    Files.writeString(trivyFile.toPath(), trivyScript)
+  private fun makeSonarFile(sonarScript: String) {
+    val sonarFile = File(projectDir, "sonar-project.properties")
+    Files.writeString(sonarFile.toPath(), sonarScript)
   }
 }
